SiamcoWeb/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-easy-pdf==0.2.0.dev1
django-heroku==0.3.1
django-utils-six==2.0
fpdf==1.7.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
reportlab==3.5.46
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
toml==0.10.1
urllib3==1.25.9
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
